ai-pipeline/image_processor/requirements.txt
# Image Processing Requirements
# ==============================

# Core
numpy>=1.24.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0

# OCR Engines
pytesseract>=0.3.10
easyocr>=1.7.0
# paddleocr>=2.7.0  # Uncomment if needed (large dependency)

# PDF Processing
pdf2image>=1.16.0
PyMuPDF>=1.23.0  # fitz

# Document Layout
# layoutparser>=0.3.4  # Uncomment for ML-based layout detection
# detectron2  # Required for layoutparser

# Language Detection
langdetect>=1.0.9

# FastAPI (for standalone service)
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Image Enhancement
scikit-image>=0.21.0

# Optional: Google Cloud Vision
# google-cloud-vision>=3.4.0
